+++ src/itasca_mcp_bridge/utils/__init__.py
@@ -6,8 +6,8 @@ from . import path_utils
 from .path_utils import path_to_llm_format
 from .file_buffer import FileBuffer, TeeBuffer
 from .response import TaskDataBuilder, build_response
-from .command_splitter import preprocess_script
-from .command_log import capture_pfc_console
+from .command_splitter import preprocess_source
+from .command_log import capture_engine_console
 from .safe_logging import GapFreeFileHandler, GapFreeStreamHandler
 
 __all__ = [
@@ -17,8 +17,8 @@ __all__ = [
     'TeeBuffer',
     'TaskDataBuilder',
     'build_response',
-    'preprocess_script',
-    'capture_pfc_console',
+    'preprocess_source',
+    'capture_engine_console',
     'GapFreeFileHandler',
     'GapFreeStreamHandler',
 ]